+++ scripts/rollback/rollback-2.12.9.sh
@@ -1,0 +1,186 @@
+#!/bin/bash
+# timestamp: 1630493702254
+
+npm dist-tag add @midwayjs/egg-layer@2.12.7 latest
+npm dist-tag add @midwayjs/egg-layer@2.12.9 beta
+tnpm dist-tag add @midwayjs/egg-layer@2.12.7 latest
+tnpm dist-tag add @midwayjs/egg-layer@2.12.9 latest
+npm dist-tag add @midwayjs/express-layer@2.12.7 latest
+npm dist-tag add @midwayjs/express-layer@2.12.9 beta
+tnpm dist-tag add @midwayjs/express-layer@2.12.7 latest
+tnpm dist-tag add @midwayjs/express-layer@2.12.9 latest
+npm dist-tag add @midwayjs/faas-typings@2.11.1 latest
+npm dist-tag add @midwayjs/faas-typings@2.12.9 beta
+tnpm dist-tag add @midwayjs/faas-typings@2.11.1 latest
+tnpm dist-tag add @midwayjs/faas-typings@2.12.9 latest
+npm dist-tag add @midwayjs/koa-layer@2.12.7 latest
+npm dist-tag add @midwayjs/koa-layer@2.12.9 beta
+tnpm dist-tag add @midwayjs/koa-layer@2.12.7 latest
+tnpm dist-tag add @midwayjs/koa-layer@2.12.9 latest
+npm dist-tag add @midwayjs/serverless-app@2.12.8 latest
+npm dist-tag add @midwayjs/serverless-app@2.12.9 beta
+tnpm dist-tag add @midwayjs/serverless-app@2.12.8 latest
+tnpm dist-tag add @midwayjs/serverless-app@2.12.9 latest
+npm dist-tag add @midwayjs/serverless-aws-starter@2.11.4 latest
+npm dist-tag add @midwayjs/serverless-aws-starter@2.12.9 beta
+tnpm dist-tag add @midwayjs/serverless-aws-starter@2.11.4 latest
+tnpm dist-tag add @midwayjs/serverless-aws-starter@2.12.9 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@2.12.7 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@2.12.9 beta
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.12.7 latest
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.12.9 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.12.7 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.12.9 beta
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.12.7 latest
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.12.9 latest
+npm dist-tag add @midwayjs/serverless-http-parser@2.11.4 latest
+npm dist-tag add @midwayjs/serverless-http-parser@2.12.9 beta
+tnpm dist-tag add @midwayjs/serverless-http-parser@2.11.4 latest
+tnpm dist-tag add @midwayjs/serverless-http-parser@2.12.9 latest
+npm dist-tag add @midwayjs/serverless-scf-starter@2.11.4 latest
+npm dist-tag add @midwayjs/serverless-scf-starter@2.12.9 beta
+tnpm dist-tag add @midwayjs/serverless-scf-starter@2.11.4 latest
+tnpm dist-tag add @midwayjs/serverless-scf-starter@2.12.9 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.12.7 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.12.9 beta
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.12.7 latest
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.12.9 latest
+npm dist-tag add @midwayjs/serverless-vercel-starter@2.12.7 latest
+npm dist-tag add @midwayjs/serverless-vercel-starter@2.12.9 beta
+tnpm dist-tag add @midwayjs/serverless-vercel-starter@2.12.7 latest
+tnpm dist-tag add @midwayjs/serverless-vercel-starter@2.12.9 latest
+npm dist-tag add @midwayjs/static-layer@2.12.7 latest
+npm dist-tag add @midwayjs/static-layer@2.12.9 beta
+tnpm dist-tag add @midwayjs/static-layer@2.12.7 latest
+tnpm dist-tag add @midwayjs/static-layer@2.12.9 latest
+npm dist-tag add @midwayjs/axios@2.12.7 latest
+npm dist-tag add @midwayjs/axios@2.12.9 beta
+tnpm dist-tag add @midwayjs/axios@2.12.7 latest
+tnpm dist-tag add @midwayjs/axios@2.12.9 latest
+npm dist-tag add @midwayjs/bootstrap@2.12.7 latest
+npm dist-tag add @midwayjs/bootstrap@2.12.9 beta
+tnpm dist-tag add @midwayjs/bootstrap@2.12.7 latest
+tnpm dist-tag add @midwayjs/bootstrap@2.12.9 latest
+npm dist-tag add @midwayjs/cache@2.12.7 latest
+npm dist-tag add @midwayjs/cache@2.12.9 beta
+tnpm dist-tag add @midwayjs/cache@2.12.7 latest
+tnpm dist-tag add @midwayjs/cache@2.12.9 latest
+npm dist-tag add @midwayjs/consul@2.12.7 latest
+npm dist-tag add @midwayjs/consul@2.12.9 beta
+tnpm dist-tag add @midwayjs/consul@2.12.7 latest
+tnpm dist-tag add @midwayjs/consul@2.12.9 latest
+npm dist-tag add @midwayjs/core@2.12.7 latest
+npm dist-tag add @midwayjs/core@2.12.9 beta
+tnpm dist-tag add @midwayjs/core@2.12.7 latest
+tnpm dist-tag add @midwayjs/core@2.12.9 latest
+npm dist-tag add @midwayjs/decorator@2.12.3 latest
+npm dist-tag add @midwayjs/decorator@2.12.9 beta
+tnpm dist-tag add @midwayjs/decorator@2.12.3 latest
+tnpm dist-tag add @midwayjs/decorator@2.12.9 latest
+npm dist-tag add @midwayjs/faas@2.12.7 latest
+npm dist-tag add @midwayjs/faas@2.12.9 beta
+tnpm dist-tag add @midwayjs/faas@2.12.7 latest
+tnpm dist-tag add @midwayjs/faas@2.12.9 latest
+npm dist-tag add @midwayjs/grpc@2.12.7 latest
+npm dist-tag add @midwayjs/grpc@2.12.9 beta
+tnpm dist-tag add @midwayjs/grpc@2.12.7 latest
+tnpm dist-tag add @midwayjs/grpc@2.12.9 latest
+npm dist-tag add midway-schedule@2.12.7 latest
+npm dist-tag add midway-schedule@2.12.9 beta
+tnpm dist-tag add midway-schedule@2.12.7 latest
+tnpm dist-tag add midway-schedule@2.12.9 latest
+npm dist-tag add midway@2.12.7 latest
+npm dist-tag add midway@2.12.9 beta
+tnpm dist-tag add midway@2.12.7 latest
+tnpm dist-tag add midway@2.12.9 latest
+npm dist-tag add @midwayjs/mock@2.12.7 latest
+npm dist-tag add @midwayjs/mock@2.12.9 beta
+tnpm dist-tag add @midwayjs/mock@2.12.7 latest
+tnpm dist-tag add @midwayjs/mock@2.12.9 latest
+npm dist-tag add @midwayjs/orm@2.12.7 latest
+npm dist-tag add @midwayjs/orm@2.12.9 beta
+tnpm dist-tag add @midwayjs/orm@2.12.7 latest
+tnpm dist-tag add @midwayjs/orm@2.12.9 latest
+npm dist-tag add @midwayjs/oss@2.12.7 latest
+npm dist-tag add @midwayjs/oss@2.12.9 beta
+tnpm dist-tag add @midwayjs/oss@2.12.7 latest
+tnpm dist-tag add @midwayjs/oss@2.12.9 latest
+npm dist-tag add @midwayjs/prometheus@2.12.7 latest
+npm dist-tag add @midwayjs/prometheus@2.12.9 beta
+tnpm dist-tag add @midwayjs/prometheus@2.12.7 latest
+tnpm dist-tag add @midwayjs/prometheus@2.12.9 latest
+npm dist-tag add @midwayjs/rabbitmq@2.12.8 latest
+npm dist-tag add @midwayjs/rabbitmq@2.12.9 beta
+tnpm dist-tag add @midwayjs/rabbitmq@2.12.8 latest
+tnpm dist-tag add @midwayjs/rabbitmq@2.12.9 latest
+npm dist-tag add @midwayjs/socketio@2.12.7 latest
+npm dist-tag add @midwayjs/socketio@2.12.9 beta
+tnpm dist-tag add @midwayjs/socketio@2.12.7 latest
+tnpm dist-tag add @midwayjs/socketio@2.12.9 latest
+npm dist-tag add @midwayjs/task@2.12.7 latest
+npm dist-tag add @midwayjs/task@2.12.9 beta
+tnpm dist-tag add @midwayjs/task@2.12.7 latest
+tnpm dist-tag add @midwayjs/task@2.12.9 latest
+npm dist-tag add @midwayjs/typegoose@2.12.7 latest
+npm dist-tag add @midwayjs/typegoose@2.12.9 beta
+tnpm dist-tag add @midwayjs/typegoose@2.12.7 latest
+tnpm dist-tag add @midwayjs/typegoose@2.12.9 latest
+npm dist-tag add @midwayjs/version@2.12.8 latest
+npm dist-tag add @midwayjs/version@2.12.9 beta
+tnpm dist-tag add @midwayjs/version@2.12.8 latest
+tnpm dist-tag add @midwayjs/version@2.12.9 latest
+npm dist-tag add @midwayjs/express@2.12.7 latest
+npm dist-tag add @midwayjs/express@2.12.9 beta
+tnpm dist-tag add @midwayjs/express@2.12.7 latest
+tnpm dist-tag add @midwayjs/express@2.12.9 latest
+npm dist-tag add @midwayjs/koa@2.12.7 latest
+npm dist-tag add @midwayjs/koa@2.12.9 beta
+tnpm dist-tag add @midwayjs/koa@2.12.7 latest
+tnpm dist-tag add @midwayjs/koa@2.12.9 latest
+npm dist-tag add @midwayjs/web@2.12.7 latest
+npm dist-tag add @midwayjs/web@2.12.9 beta
+tnpm dist-tag add @midwayjs/web@2.12.7 latest
+tnpm dist-tag add @midwayjs/web@2.12.9 latest
+npm dist-tag add @midwayjs/ws@2.12.7 latest
+npm dist-tag add @midwayjs/ws@2.12.9 beta
+tnpm dist-tag add @midwayjs/ws@2.12.7 latest
+tnpm dist-tag add @midwayjs/ws@2.12.9 latest
+
+# Changes:
+
+#  - @midwayjs/egg-layer: 2.12.7 => 2.12.9
+#  - @midwayjs/express-layer: 2.12.7 => 2.12.9
+#  - @midwayjs/faas-typings: 2.11.1 => 2.12.9
+#  - @midwayjs/koa-layer: 2.12.7 => 2.12.9
+#  - @midwayjs/serverless-app: 2.12.8 => 2.12.9
+#  - @midwayjs/serverless-aws-starter: 2.11.4 => 2.12.9
+#  - @midwayjs/serverless-fc-starter: 2.12.7 => 2.12.9
+#  - @midwayjs/serverless-fc-trigger: 2.12.7 => 2.12.9
+#  - @midwayjs/serverless-http-parser: 2.11.4 => 2.12.9
+#  - @midwayjs/serverless-scf-starter: 2.11.4 => 2.12.9
+#  - @midwayjs/serverless-scf-trigger: 2.12.7 => 2.12.9
+#  - @midwayjs/serverless-vercel-starter: 2.12.7 => 2.12.9
+#  - @midwayjs/static-layer: 2.12.7 => 2.12.9
+#  - @midwayjs/axios: 2.12.7 => 2.12.9
+#  - @midwayjs/bootstrap: 2.12.7 => 2.12.9
+#  - @midwayjs/cache: 2.12.7 => 2.12.9
+#  - @midwayjs/consul: 2.12.7 => 2.12.9
+#  - @midwayjs/core: 2.12.7 => 2.12.9
+#  - @midwayjs/decorator: 2.12.3 => 2.12.9
+#  - @midwayjs/faas: 2.12.7 => 2.12.9
+#  - @midwayjs/grpc: 2.12.7 => 2.12.9
+#  - midway-schedule: 2.12.7 => 2.12.9
+#  - midway: 2.12.7 => 2.12.9
+#  - @midwayjs/mock: 2.12.7 => 2.12.9
+#  - @midwayjs/orm: 2.12.7 => 2.12.9
+#  - @midwayjs/oss: 2.12.7 => 2.12.9
+#  - @midwayjs/prometheus: 2.12.7 => 2.12.9
+#  - @midwayjs/rabbitmq: 2.12.8 => 2.12.9
+#  - @midwayjs/socketio: 2.12.7 => 2.12.9
+#  - @midwayjs/task: 2.12.7 => 2.12.9
+#  - @midwayjs/typegoose: 2.12.7 => 2.12.9
+#  - @midwayjs/version: 2.12.8 => 2.12.9
+#  - @midwayjs/express: 2.12.7 => 2.12.9
+#  - @midwayjs/koa: 2.12.7 => 2.12.9
+#  - @midwayjs/web: 2.12.7 => 2.12.9
+#  - @midwayjs/ws: 2.12.7 => 2.12.9