+++ packages-serverless/serverless-http-parser/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.12.9](https://github.com/midwayjs/midway-faas/compare/v2.12.8...v2.12.9) (2021-09-01)
+
+
+### Features
+
+* add ctx.throw for serverless app ([#1262](https://github.com/midwayjs/midway-faas/issues/1262)) ([da58bf6](https://github.com/midwayjs/midway-faas/commit/da58bf647f23077fc4142f6454cd6c40d8a82e96))
+
+
+
+
+
 ## [2.11.4](https://github.com/midwayjs/midway-faas/compare/v2.11.3...v2.11.4) (2021-07-06)